create components FetchGameInfo for Gamespages
